# shackServer/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Autoshack API",
    "description": "API for the Autoshack project"
  },
  "servers": [
    {
      "url": "http://localhost:1783",
      "description": ""
    }
  ],
...
  "paths": {
...
    "/schedule/": {
      "get": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/schedule/update": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "start_hour": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  },
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schedule/add": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "start_hour": {
                    "example": "any"
                  },
                  "duration": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/schedule/delete": {
      "post": {
        "tags": [
          "Schedule"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}